feat(ui-dsc): add ShoppingCartList component
The ShoppingCartList component is added to the ui-dsc package. This component is responsible for rendering the list of items in the shopping cart. It takes an array of cart items and a deleteItem callback function as props. The component displays the list of items using the ShoppingCartItem component and also includes a ShoppingCartTotalCard component to display the total price of the items in the cart. Additionally, if the cart is not full, a ContinueShoppingCard component is displayed.