feat(cartState.ts): add removeItemAtom to remove an item from the strainItemsAtom array
The removeItemAtom is added to the cartState.ts file to provide a way to remove an item from the strainItemsAtom array. This atom takes an ID as a parameter and filters the strainItemsAtom array to remove the item with the matching ID. This allows for easy removal of items from the cart.