feat(ui-dsc): add isFull utility function
The isFull utility function is added to the ui-dsc package. This function takes an array as input and returns true if the length of the array is equal to 12, indicating that the array is full. This utility function can be used to check if an array has reached its maximum capacity.