feat(functional.tsx): add renderCartTotal function
The renderCartTotal function is added to the functional.tsx file. This function renders a CartTotalRowV2 component that displays the total value of items in the cart. It takes in the strainItems and plasmidItems from the Cart object and renders the total value. This function is exported along with the existing renderStrainTotal, renderPlasmidTotal, and renderStrainAndPlasmidTotals functions.