feat(ui-dsc): add ShippingPageRHF component for entering shipping information
The ShippingPageRHF component is added to the ui-dsc package. This component is responsible for displaying a form where users can enter their shipping information. It uses react-hook-form and Yup for form validation. The form data is stored in the formData state and can be updated using the setFormData function. The nextStep function is called when the form is submitted.