feat(OrderForm.tsx): add OrderForm component
The OrderForm component is added to the project. It imports necessary dependencies such as "ts-pattern", "@dictybase/ui-dsc", and "jotai". The component uses the "orderStepAtom" atom from the state file to determine the current step of the order process. Based on the value of "orderStep", the component renders different pages such as ShippingPage, PaymentPage, or SubmitPage. If the value of "orderStep" does not match any of the defined steps, an error message is displayed.