chore(vitest.config.ts): update setupFiles path in vitest.config.ts files
The setupFiles path in the vitest.config.ts files for multiple packages has been updated to "./setup.ts" instead of "./src/__tests__/setup.ts".
This is because the latest version of vitest considers all files in the __tests__ directory to be test files and tries to run them as such, which would cause a test failure. Moving setup.ts out of this directory fixes this issue.