feat(ui-dsc): add ShippingMethodRadioGroup component
The ShippingMethodRadioGroup component is added to the ui-dsc package. This component provides a set of radio buttons for selecting a shipping method. It allows the user to choose between different courier options (DHL, FedEx, UPS) or to select the option to send a prepaid shipping label. The component also includes functions to handle changes in the shipping account number and prepaid notice.