feat(ui-dsc): add ShippingMethodPrepaidNotice component
The ShippingMethodPrepaidNotice component is added to the ui-dsc package. This component displays a notice for sending a prepaid shipping label and includes a link to the email address where the label should be sent. This component improves the user experience by providing clear instructions for sending a prepaid shipping label.