feat(BackButton.tsx): add BackButton component
The BackButton component is added to the project. It displays a button with a back arrow icon and the text "Back". When clicked, it decreases the order step by 1 using the `setOrderStep` function from the `orderStepAtom` state. This component is used on each page of the order form to allow users to navigate back to the previous step.