feat(SubmitPage.tsx): add SubmitPage component
The SubmitPage component is added to the project. This component is the final page that the user sees before submitting the order. It displays the order summary, including the shipping and payment form data, as well as the items in the cart. It also includes a BackButton component for navigating back to the previous page. The SubmitButton component is currently commented out and not functional.