feat(ui-dsc): add getListValues utility module
This commit adds a new utility module called `getListValues.ts` to the `ui-dsc` package. This module contains two functions: `getShippingValues` and `getPaymentValues`.
The `getShippingValues` function takes in a `FormData` object and generates an array of strings to display for the shipping section. It extracts specific fields from the `FormData` object and formats them accordingly.
The `getPaymentValues` function also takes in a `FormData` object and generates an array of strings to display for the payment section. It extracts specific fields from the `FormData` object and formats them accordingly.

These utility functions will be used to generate the values to display in the UI for the shipping and payment sections of a form.