feat(ui-dsc): add SubmitError component for displaying form submission errors
The SubmitError component is added to the ui-dsc package. This component is responsible for displaying an error message when there is an error in form submission. It includes a Material-UI Alert component with an error severity and a Typography component for the error message. The error message provides information about the error and includes a link to contact the support team via email.