feat(ui-dsc): add DetailsHeaderCopyIcon component
The DetailsHeaderCopyIcon component is added to the ui-dsc package. This component displays a copy icon in the header of every stock details page. When clicked, it copies the stock ID to the clipboard and displays a success message. This component improves the user experience by providing an easy way to copy the stock ID.