feat(characterConverter): add characterConverter utility function
The characterConverter utility function is added to the `utils` directory. This function is used to replace any HTML Symbol Entities with their corresponding characters. It uses a regular expression to convert numeric entities and manually replaces the named entity "&gamma;" with the character "γ". This utility function improves the handling of HTML entities in the application.