feat(ui-dsc): add DetailsListItem component
The DetailsListItem component is added to the catalog directory in the ui-dsc package. This component is used to display a single row of data for a stock item. It takes in a title and content as properties, and renders them in a ListItem with Typography components for the title and content. The content can be a string, JSX element, or an array of JSX elements. The component also uses a characterConverter function to convert any HTML entities in the content.