feat(ui-dsc): add AddToCartDialogContent component
The AddToCartDialogContent component is added to the catalog package in the ui-dsc package. This component is responsible for rendering the main content of the add to cart dialog box. It receives an array of stock data as a prop and maps over the data to render the necessary information for each item.