feat(ui-dsc): add AddToCartDialogActions component
The AddToCartDialogActions component is added to the catalog package in the ui-dsc package. This component is responsible for displaying the action buttons at the bottom of the cart dialog box. It includes buttons for continuing shopping and viewing the cart. The component also accepts a handleClose function as a prop, which is called when closing the dialog box.