feat(ui-dsc): add LinkTag component
The LinkTag component is added to the catalog directory in the ui-dsc package. This component is a tag-style display for showing links. It takes in two properties: item (the item to link to) and route (the subroute to use, such as publication or gene). The component renders a Button component with a link to the specified route and item. It also includes an external link icon from the FontAwesomeIcon component.