feat(ui-dsc): add phenotypeStyles.tsx to catalog folder
The phenotypeStyles.tsx file is added to the catalog folder in the ui-dsc package. This file contains the makeStyles function from the @material-ui/core/styles library, along with styles for various elements such as paper, row, item, listHeaders, list, and button. These styles are used for the phenotype catalog component in the UI.