feat(ui-dsc): add StrainPhenotypeList component
The StrainPhenotypeList component is added to the ui-dsc package. This component provides a list of phenotypes for a given strain. It uses the Material-UI Paper component for styling and renders a header component (StrainPhenotypeListHeader) and a list of phenotype items (StrainPhenotypeListItem). The phenotypes are sorted alphabetically by phenotype name using the fp-ts library.