feat(ui-dsc): add useGenotypes hook for separating genotype data
The useGenotypes hook is added to the ui-dsc package. This hook takes in a string of genotypes and separates them into drug resistances and non-drug resistances. The drug resistances are determined based on a predefined set of resistance markers. This hook provides a convenient way to handle and categorize genotype data in the UI.